chore: satisfy lint for CI\n\n- Remove unused imports/vars\n- Refactor month context to avoid react-refresh rule (move hook/utils)\n- Replace empty catch blocks with no-op\n- Trim unused parameters and dead code in Incomes/Budget/Plan
