+++ pf-dashboard/src/lib/monthContext.jsx
@@ -1,7 +1,6 @@
-import { createContext, useContext, useMemo, useState, useEffect } from 'react'
+import { useMemo, useState, useEffect } from 'react'
 import { format } from 'date-fns'
-
-const MonthContext = createContext(null)
+import { MonthContext } from './monthStore.js'
 
 export function MonthProvider({ children }) {
   const now = new Date()
@@ -15,24 +14,4 @@ export function MonthProvider({ children }) {
   const value = useMemo(() => ({ monthKey, setMonthKey }), [monthKey])
   return <MonthContext.Provider value={value}>{children}</MonthContext.Provider>
 }
-
-export function useMonth() {
-  const ctx = useContext(MonthContext)
-  if (!ctx) throw new Error('useMonth must be used within MonthProvider')
-  return ctx
-}
-
-export function nextMonthKey(monthKey) {
-  const [y, m] = monthKey.split('-').map(Number)
-  const d = new Date(y, m - 1, 1)
-  d.setMonth(d.getMonth() + 1)
-  return format(d, 'yyyy-MM')
-}
-
-export function prevMonthKey(monthKey) {
-  const [y, m] = monthKey.split('-').map(Number)
-  const d = new Date(y, m - 1, 1)
-  d.setMonth(d.getMonth() - 1)
-  return format(d, 'yyyy-MM')
-}
-
+// month navigation helpers moved to lib/monthUtils.js; hook moved to lib/useMonth.js
