+++ pf-dashboard/src/lib/monthStore.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+
+export const MonthContext = createContext(null)
+
