+++ pf-dashboard/src/lib/monthUtils.js
@@ -1,0 +1,16 @@
+import { format } from 'date-fns'
+
+export function nextMonthKey(monthKey) {
+  const [y, m] = monthKey.split('-').map(Number)
+  const d = new Date(y, m - 1, 1)
+  d.setMonth(d.getMonth() + 1)
+  return format(d, 'yyyy-MM')
+}
+
+export function prevMonthKey(monthKey) {
+  const [y, m] = monthKey.split('-').map(Number)
+  const d = new Date(y, m - 1, 1)
+  d.setMonth(d.getMonth() - 1)
+  return format(d, 'yyyy-MM')
+}
+
